Change main to class Game
